Stop Sidebar "growing" when opened a second time. - done.
Some work still to be done with themes that have a large font with relationship name.
Would also like to dynamically adjust Navigator height based on number of individuals in list.
Not sure this is feasible yet.
However, next step is to add a "pin" function to Sidebar, to keep it open when changing indis.