Pedigree - Hourglass Scroll Bars
Unfortunately, the patch I submitted for issue 1692 did not do the job.
Ideal, an overflow-y: visible; would be the optimal solution. However,
this is not possible due to use of overflow-x (forces visible to become
auto). As such, a vertical scroll bar is visible when not really needed
unless you mouse over the pedigree popup.