improve edit control for SOUR/DATA/EVEN
This enhances the changes made a while back (#1377 etc):
The EVEN field is text such as "2 EVEN CHR, MARR, BURI", but we should provide a proper edit control for this. Events that may be selected here are configurable in the preferences.
The implementation uses a custom select2 control in order to preserve the order of elements.