Report facilities expanded
The app/Report tool creates reports from the collected gedcom data. The current versio has some
limitation and shortcomings that this update addresses. A detailed description is incuded as a
pdf file, and in short the changes are:

Report tries to place text on an exact position on a page. But many pitfalls on the way. Counting lines
in a textbox is not enough, the lines may be set in different font sizes; converting lines to points
in HTML uses a much too high number, which results in a lot of white space, especially for long reports
(many pages). The PDF reports have problems to predict when the tool TcPdf will break a line or add a page.
This applies to the files HtmlRenderer.php, ReportHtmlTextbox.php, ReportHtmlText.php, PdfRenderer.php,
ReportPdfTextBox.php and ReportPdfText.php,

The file ReportParserGenerate.php is responsible for the interpretation of the xml code that defines
the layout and contents of the report. A number of addition have been introduced.
(1) GetPersonName has got a new attribute, select=..., that allows the report to present the last name
used or a combination of latest and surname at birth. Another attribute, fam_relation="1", will add
the relation the current person to the start person of a report, e.g. mmf which means mother's
mother's father.
(2) The value that is returned by GedcomValue can be a very long string. An attribute truncate=... already
exists but has been two new possibilities for dates: "Y" will give only the year and "d" gives the date
with the three character short for the month, "dd mmm yyyy".
(3) The calculation of the family relation that can be give with the name requires a substatial amount
of code.
(4) The xml keyword Facts returns all facts and events of the current object (person, family) in the
order they appear. But in a report one wants chronological order and that has been implemented. A special
addition is that a temporary gedcom event is introduced in the process, _SP_DEAT, which when a spouse
has died, if applicable. This is used only during calculations, not added to the Gedcom database or file.
This (hidden) gedcom value must be added to the gedcom.php file.
(A similar presentation is given on the main page for each person.)
(5) The keyword SetVar has been expanded with new possibilities for the attribute value=.... The value can
now specify "@$xxx" which means the the value of xxx is interpreted as the name of the varible to select.
Other new values are "@format" that responds PDF or HTML depending on whichever form of report is selected.
"@$dump" can give some help in debugging new or modified reports.
(6) Sometimes ancestors have married a relative. In pedigree charts these persons were given as separate
persons, although with tha same name, birthdate, et c. But now such persons are detected and the second
occurrance is given as "same person as ...".
(7) Some errors are corrected.