Updated standard Reports
Some of the reports supplied with webtrees relies on the current app/Report tool. The layout can be
slightly changed with the suggested new app/report. A couple of reports try to give the result as tables.
The changes here take care of that. These new  xml files work well with the old app/Report xml tool.