On branch main Your branch is up to date with 'origin/main'.
Revert "UUID References in notes" since the pull request was not
accepted

You are currently reverting commit 06d797c916fc4fb2732cd685398b6a7e35898512.

Reasons are:

---

fisharebest commented on Apr 24
This does a full-text search of the database for each UID.

This is likely to be slow on large sites.

We can do this for XREFs because these are indexed.

---

fisharebest commented on May 13
Changing the database structure needs a lot of careful thought.

It also needs some sort of data migration so that existing records are stored.

This is already planned - I want to change webtrees from being "record based" (level 0 GEDCOM) to being "fact based" (level 1 GEDCOM). This will have an index on the fact type (e.g. UID) and the fact value. But a lot of work is needed, and so won't be ready for a long time.

---